+++ inst/examples/box_elements/user_messages.R
@@ -1,0 +1,24 @@
+user_messages <- 'boxPlus(
+  width = 12,
+  "Box with messages",
+  solidheader = TRUE,
+  status = "warning",
+  userMessages(
+    width = 12,
+    status = "success",
+    userMessage(
+      author = "Alexander Pierce",
+      date = "20 Jan 2:00 pm",
+      src = "https://adminlte.io/themes/AdminLTE/dist/img/user1-128x128.jpg",
+      side = NULL,
+      "Is this template really for free? That iss unbelievable!"
+    ),
+    userMessage(
+      author = "Sarah Bullock",
+      date = "23 Jan 2:05 pm",
+      src = "https://adminlte.io/themes/AdminLTE/dist/img/user3-128x128.jpg",
+      side = "right",
+      "You better believe it!"
+    )
+  )
+)'