+++ NEWS.md
@@ -1,8 +1,10 @@
 # shinydashboardPlus 0.7.5
 
+## Minor Changes
+- Fix #40: add sidebar_icon argument to `boxPlus()`. Thanks @jmw86069 for the suggestion
+
 ## Bug Fixes
-- Fix #36: make sure that a `carousel()` item is displayed when it is generated via a shiny Output function. 
-Thanks @daattali for the report
+- Fix #36: make sure that a `carousel()` item is displayed when it is generated via a shiny Output function. Thanks @daattali for the report
 
 
 # shinydashboardPlus 0.7.0